chore(cjs-module-analyzer): Clarify package name
