+++ packages/cjs-module-analyzer/index.js
@@ -1,0 +1,1 @@
+export { analyzeCommonJS } from './src/cjs-module-analyzer.js';
