+++ packages/cjs-module-analyzer/jsconfig.json
