+++ packages/cjs-module-analyzer/src/cjs-module-analyzer.js
