--- packages/lexer/index.d.ts
@@ -1,1 +1,0 @@
-export type * from './src/main.js';
