--- packages/lexer/src/main.js
@@ -1,1 +1,0 @@
-export { analyzeCommonJS } from './lexer.js';
